Move test and remove duplicate fixture
